修改包结构
