+++ springboot-sample/pom.xml
@@ -9,7 +9,7 @@
 	<packaging>jar</packaging>
 
 	<name>springboot-sample</name>
-	<description>Demo project for Spring Boot</description>
+	<description>springboot-sample</description>
 
 	<parent>
 		<groupId>com.lxg</groupId>
@@ -60,6 +60,14 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-thymeleaf</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-autoconfigure</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-aop</artifactId>
+		</dependency>
 
 		<!-- swagger2 -->
 		<dependency>
@@ -72,7 +80,6 @@
 			<artifactId>springfox-swagger-ui</artifactId>
 			<version>2.2.2</version>
 		</dependency>
-
 	</dependencies>
