--- springboot-sample/src/main/java/com/lxg/DataSourceConfig.java
@@ -1,48 +1,0 @@
-//package com.lxg;
-//
-//import org.springframework.beans.factory.annotation.Qualifier;
-//import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
-//import org.springframework.boot.context.properties.ConfigurationProperties;
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.context.annotation.Primary;
-//import org.springframework.jdbc.core.JdbcTemplate;
-//
-//import javax.sql.DataSource;
-//
-///**
-// * Created by 刘雪岗 on 2017/1/4.
-// */
-//
-///**
-// * 多数据源配置
-// */
-//@Configuration
-//public class DataSourceConfig {
-//    @Bean(name = "primaryDataSource")
-//    @Qualifier("primaryDataSource")
-//    @ConfigurationProperties(prefix="spring.datasource.primary")
-//    public DataSource primaryDataSource() {
-//        return DataSourceBuilder.create().build();
-//    }
-//
-//    @Bean(name = "secondaryDataSource")
-//    @Qualifier("secondaryDataSource")
-//    @Primary
-//    @ConfigurationProperties(prefix="spring.datasource.secondary")
-//    public DataSource secondaryDataSource() {
-//        return DataSourceBuilder.create().build();
-//    }
-//
-//    @Bean(name = "primaryJdbcTemplate")
-//    public JdbcTemplate primaryJdbcTemplate(
-//            @Qualifier("primaryDataSource") DataSource dataSource) {
-//        return new JdbcTemplate(dataSource);
-//    }
-//
-//    @Bean(name = "secondaryJdbcTemplate")
-//    public JdbcTemplate secondaryJdbcTemplate(
-//            @Qualifier("secondaryDataSource") DataSource dataSource) {
-//        return new JdbcTemplate(dataSource);
-//    }
-//}
