+++ springboot-sample/src/main/java/com/lxg/sample/SpringBootDemoApplication.java
@@ -1,13 +1,18 @@
-package com.lxg;
+package com.lxg.sample;
 
+import com.lxg.sample.service.UserService;
+import com.lxg.service.UserServiceImpl;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
 
-/**
- * 主程序
- */
 @SpringBootApplication
 public class SpringBootDemoApplication {
+
+	@Bean
+	public UserService getUserService(){
+		return new UserServiceImpl();
+	}
 
 	public static void main(String[] args) {
 		SpringApplication.run(SpringBootDemoApplication.class, args);
