+++ springboot-sample/src/main/java/com/lxg/sample/base/ErrorInfo.java
@@ -1,10 +1,7 @@
-package com.lxg.entity;
+package com.lxg.sample.base;
 
 /**
  * Created by 刘雪岗 on 2017/1/4.
- */
-
-/**
  * 异常处理实体类
  * @param <T>
  */
