+++ springboot-sample/src/main/java/com/lxg/sample/config/DataSourceConfig.java
@@ -1,0 +1,55 @@
+//package com.lxg.sample.config;
+//
+//import org.springframework.beans.factory.annotation.Qualifier;
+//import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+//import org.springframework.boot.context.properties.ConfigurationProperties;
+//import org.springframework.boot.jdbc.DataSourceBuilder;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.jdbc.core.JdbcTemplate;
+//
+//import javax.sql.DataSource;
+//
+//@Configuration
+//@EnableAutoConfiguration
+//public class DataSourceConfig {
+//
+//    @Bean("dataSource")
+//    public DataSource dataSource() {
+//        return DataSourceBuilder.create().build();
+//    }
+//
+//    @Bean(name = "jdbcTemplate")
+//    public JdbcTemplate jdbcTemplate(
+//            @Qualifier("dataSource") DataSource dataSource) {
+//        return new JdbcTemplate(dataSource);
+//    }
+//
+//
+////    @Bean(name = "primaryDataSource")
+////    @Qualifier("primaryDataSource")
+////    @ConfigurationProperties(prefix="spring.datasource.primary")
+////    public DataSource primaryDataSource() {
+////        return DataSourceBuilder.create().build();
+////    }
+////
+////    @Bean(name = "secondaryDataSource")
+////    @Qualifier("secondaryDataSource")
+////    @Primary
+////    @ConfigurationProperties(prefix="spring.datasource.secondary")
+////    public DataSource secondaryDataSource() {
+////        return DataSourceBuilder.create().build();
+////    }
+////
+////    @Bean(name = "primaryJdbcTemplate")
+////    public JdbcTemplate primaryJdbcTemplate(
+////            @Qualifier("primaryDataSource") DataSource dataSource) {
+////        return new JdbcTemplate(dataSource);
+////    }
+////
+////    @Bean(name = "secondaryJdbcTemplate")
+////    public JdbcTemplate secondaryJdbcTemplate(
+////            @Qualifier("secondaryDataSource") DataSource dataSource) {
+////        return new JdbcTemplate(dataSource);
+////    }
+//}
