+++ springboot-sample/src/main/java/com/lxg/sample/exception/MyException.java
@@ -1,10 +1,7 @@
-package com.lxg.exception;
+package com.lxg.sample.exception;
 
 /**
  * Created by 刘雪岗 on 2017/1/4.
- */
-
-/**
  * 自定义异常类
  */
 public class MyException extends Exception {
