+++ springboot-sample/src/main/resources/application-dev.properties
@@ -1,0 +1,9 @@
+# mysql连接
+spring.datasource.url=jdbc:mysql://140.143.250.99:3306/lxg_test
+spring.datasource.username=root
+spring.datasource.password=lxgmysql
+spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+
+server.port=8082
+server.tomcat.uri-encoding=utf-8
+server.servlet.context-path=/sample
