+++ springboot-sample/src/main/resources/application.properties
@@ -1,17 +1,1 @@
-
-# mysql连接
-spring.datasource.url=jdbc:mysql://localhost:3306/test
-spring.datasource.username=root
-spring.datasource.password=root
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-
-## 多数据源配置 1
-#spring.datasource.primary.url=jdbc:mysql://localhost:3306/test1
-#spring.datasource.primary.username=root
-#spring.datasource.primary.password=root
-#spring.datasource.primary.driver-class-name=com.mysql.jdbc.Driver
-## 多数据源配置 2
-#spring.datasource.secondary.url=jdbc:mysql://localhost:3306/test2
-#spring.datasource.secondary.username=root
-#spring.datasource.secondary.password=root
-#spring.datasource.secondary.driver-class-name=com.mysql.jdbc.Driver
+spring.profiles.active=dev
